leam-base.commons-lang.mid-411.idx-40259.22.mutant

+++ commons-lang/src/main/java/org/apache/commons/lang3/mutable/MutableFloat.java
@@ -269,7 +269,7 @@ public class MutableFloat extends Number implements Comparable<MutableFloat>, Mu
     public float getAndAdd(final float operand) {
         final float last = value;
         this.value += operand;
-        return last;
+        return (float) last;
     }
 
     /**
